@yuryalencar: Provavelmente último commit.
Próxima etapa: Correção de no máximo alguns últimos bugs.

+++ JavaApplication19/src/org/unipampa/rp2/tiposimoveis/Chacara.java
@@ -121,7 +121,7 @@ public class Chacara extends Imovel implements Cloneable{
     /**
      * @param areaConstruida the areaConstruida to set
      */
-    public void setAreaConstruida(float areaConstruida) {
+    public void setAreaConstruida(double areaConstruida) {
         this.areaConstruida = areaConstruida;
     }
